LMSUZH-144: Cleanup process after batch import uses semester of import, not current semester -> avoid problems at change of semester
Reverted that only courses of most recent import are synchronized (i.e.  reverted previous commit)

+++ src/main/java/ch/uzh/campus/connectors/CampusImportJobInterceptor.java
@@ -2,6 +2,7 @@ package ch.uzh.campus.connectors;
 
 import ch.uzh.campus.data.DaoManager;
 import ch.uzh.campus.data.LecturerIdCourseId;
+import ch.uzh.campus.data.Semester;
 import ch.uzh.campus.data.StudentIdCourseId;
 import org.olat.core.commons.persistence.DB;
 import org.olat.core.logging.OLog;
@@ -46,9 +47,15 @@ public class CampusImportJobInterceptor implements JobExecutionListener {
 			return;
 		}
 
+		Semester semesterOfCurrentImportProcess = daoManager.getSemesterOfMostRecentCourseImport();
+		if (semesterOfCurrentImportProcess == null) {
+			LOG.warn("Current semester of import process could not be determined. Not updated data cannot be removed. Synchronization may not work properly!");
+			return;
+		}
+
 		int lecturerCoursesToBeRemoved = daoManager.deleteAllLCBookingTooFarInThePast(jobExecution.getStartTime());
 		dbInstance.intermediateCommit();
-		List<LecturerIdCourseId> lecturerIdCourseIdsToBeRemoved = daoManager.getAllNotUpdatedLCBookingOfCurrentSemester(jobExecution.getStartTime());
+		List<LecturerIdCourseId> lecturerIdCourseIdsToBeRemoved = daoManager.getAllNotUpdatedLCBookingOfCurrentImportProcess(jobExecution.getStartTime(), semesterOfCurrentImportProcess);
 		lecturerCoursesToBeRemoved += lecturerIdCourseIdsToBeRemoved.size();
 		LOG.info("LECTURER_COURSES TO BE REMOVED ["  + lecturerCoursesToBeRemoved + "]");
 		if (!lecturerIdCourseIdsToBeRemoved.isEmpty()) {
@@ -58,7 +65,7 @@ public class CampusImportJobInterceptor implements JobExecutionListener {
 
 		int studentCoursesToBeRemoved = daoManager.deleteAllSCBookingTooFarInThePast(jobExecution.getStartTime());
 		dbInstance.intermediateCommit();
-		List<StudentIdCourseId> studentIdCourseIdsToBeRemoved = daoManager.getAllNotUpdatedSCBookingOfCurrentSemester(jobExecution.getStartTime());
+		List<StudentIdCourseId> studentIdCourseIdsToBeRemoved = daoManager.getAllNotUpdatedSCBookingOfCurrentImportProcess(jobExecution.getStartTime(), semesterOfCurrentImportProcess);
 		studentCoursesToBeRemoved += studentIdCourseIdsToBeRemoved.size();
 		LOG.info("STUDENT_COURSES TO BE REMOVED [" + studentCoursesToBeRemoved + "]");
 		if (!studentIdCourseIdsToBeRemoved.isEmpty()) {
